finishing up syncing gradescope assignments
