+++ app/controllers/courses_controller.rb
@@ -22,9 +22,9 @@ class CoursesController < ApplicationController
     @side_nav = 'show'
     return redirect_to courses_path, alert: 'Course not found.' unless @course
 
-    course_to_canvas      = @course.course_to_lms(1)
-    course_to_gradescope  = @course.course_to_lms(2)
-    return redirect_to courses_path, alert: 'No LMS data found for this course.' unless course_to_canvas || course_to_gradescope
+    course_to_canvas = @course.course_to_lms(1)
+    course_to_gradescope = @course.course_to_lms(2) if @course.course_settings.enable_gradescope
+    return redirect_to courses_path, alert: 'No LMS data found for this course.' unless course_to_canvas
 
     if @role == 'student'
       course_settings = @course.course_settings
