+++ app/javascript/controllers/assignment_controller.js
@@ -1,4 +1,7 @@
 import { Controller } from "@hotwired/stimulus"
+import DataTable from "datatables.net-bs5";
+import "datatables.net-responsive";
+import "datatables.net-responsive-bs5";
 
 // Connects to data-controller="assignment"
 export default class extends Controller {
@@ -9,6 +12,16 @@ export default class extends Controller {
     this.checkboxTargets.forEach((checkbox) => {
       checkbox.addEventListener("change", (event) => this.toggleAssignment(event, checkbox))
     })
+
+    if (!DataTable.isDataTable('#assignments-table')) {
+			new DataTable('#assignments-table', {
+				paging: true,
+				searching: true,
+				ordering: true,
+				info: true,
+				responsive: true,
+			});
+		}
   }
 
   async toggleAssignment(event, checkbox) {
