+++ app/models/course_settings.rb
@@ -2,6 +2,7 @@ class CourseSettings < ApplicationRecord
   belongs_to :course
 
   before_save :ensure_system_user_for_auto_approval
+  validate :gradescope_url_is_valid, if: :enable_gradescope?
 
   private
 
@@ -10,5 +11,11 @@ class CourseSettings < ApplicationRecord
     return unless enable_extensions && auto_approve_days.to_i.positive?
 
     SystemUserService.ensure_auto_approval_user_exists
+  end
+
+  def gradescope_url_is_valid
+    return if gradescope_course_url.match?(%r{\Ahttps://(www\.)?gradescope\.com/courses/\d{6}\z})
+
+    errors.add(:gradescope_course_url, 'must be a valid Gradescope course URL like https://gradescope.com/courses/123456')
   end
 end
