--- lib/gradescope/errors.rb
@@ -1,6 +1,0 @@
-module Gradescope
-  class Error < StandardError; end
-  class AuthenticationError < Error; end
-  class NotFoundError < Error; end
-  class RequestError < Error; end
-end
