--- lib/gradescope/version.rb
@@ -1,3 +1,0 @@
-module Gradescope
-  VERSION = '0.1.0'.freeze
-end
